% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\name{calendar}
\alias{calendar}
\title{Calibration of radiocarbon dates}
\arguments{
\item{chrono_data}{Name of text file read from working directory.
File must contain three or four columns without row names or column headings. Column 1 = non-calibrated dates, column 2 = non-calibrated errors, column 3 = date identifiers, column 4 (optional) = name of calibration curve per date.
Unlimited number of dates (rows).
All dates must be Before Present, where Present = 1950.
See details.}

\item{cal_curve}{Character. Sets unique calibration curve for calibrating all radiocarbon dates.
Default \code{cal_curve = "intcal20"}.
See details.}

\item{upper14C}{Numeric. Sets the upper temporal boundary for radiocarbon calibrations.
Default \code{upper14C = 55000} for calibration curves \code{"intcal20"}, \code{"shcal20"} or \code{"marine20"}.
Use \code{upper14C = 50000} for calibration curves \code{"intcal13"}, \code{"shcal13"}, \code{"marine13"}, \code{"intcal13nhpine16"} or \code{"shcal13shkauri16"}.
Dates from 1951 to present (post-bomb) not supported.
See details.}

\item{cal_ci}{Character. Determines width of confidence interval of calibrated mean and median.
Default \code{cal_ci = "sd1"} for 68.3 percent of the distribution of the calibrated mean (plus/minus 1 standard deviation) or 68.3 percent quartile range of the median.
\code{cal_ci = "sd2"} for 95.4 percent (plus/minus 1 standard deviation).
See details.}

\item{cal_save}{Character. Determines if calibrated outputs are saved as text file to working directory.
Default \code{cal_save = "intercepts"} saves calibrated intercepts (weighted mean, median, mode).
\code{cal_save = "full"} saves PDF moments (mean, variance, skewness, kurtosis), statistically supported modes, and calibrated intercepts.
\code{cal_save = "none"} cancels saving of outputs.
See details.}
}
\value{
Returns a list containing four elements:
\itemize{
\item {\code{calibration_arguments}} {Data.frame with 2 columns.
 Column 1 = arguments in function \code{calendar}, Column 2 = values selected by user.}\cr
 \item {\code{cal_moments}} {Data.table with 6 columns and as many rows as dates in the input data.
 Column 1 = date identifiers, Column 2 = PDF mean, Column 3 = PDF variance, Column 4 = PDF skewness, Column 5 = PDF kurtosis, Column = calibration curves.}\cr
 \item {\code{cal_modality}} {Data.table with at least 5 columns and as many rows as dates in the input data.
 Column 1 = date identifiers, Column 2 = number of PDF modes at +/- 1SD, Column 3 = number of modes at +/- 2SD, last column = calibration curve.
 Remaining intermediate columns show upper and lower boundaries of each statistically supported mode in the PDF of each date.}\cr
 \item {\code{cal_incercepts}} {Data.table with 10 columns and as many rows as dates in the input data.
 Column 1 = date identifiers, Column 2 = non-calibrated date, Column 3 = non-calibrated error,
 Column 4 = calibrated weighted mean, Column 5 = calibrated weighted error, Column 6 = calibration curve,
 Column 7 = calibrated median (50 percent quartile), Column 8 = calibrated lower quartile, Column 9 = calibrated upper quartile, Column 10 = calibrated mode.}
}
}
\description{
Calibrate radiocarbon dates into point-estimate intercepts (mean, median, mode) in calendar years and describe the shape of calibrated probability density functions (PDF).
}
\details{
The concentration of the radioactive isotope carbon 14 (14C) in an organism starts decaying at a near-constant rate (Libby half-life = 5568 years) after death. The date of death can be then estimated by measuring the amount of 14C in the tissue of the organism in question as radioactivity (beta counting) or number of 14C atoms (accelerator mass spectrometry). Organic materials older some 48 thousand years retain negligible 14C content today so, currently, radiocarbon cannot be used for dating samples beyond that time interval.

Critically, 14C decay rates are non-linear so raw 14C dates in years (as obtained from dating facilities) must be adjusted to 14C decay trends based on calibration curves (as obtained from natural archives including corals, sediments, speleothems and tree rings). Calibration curves differ depending on the region (Northern versus Southern Hemisphere) or environment (terrestrial versus marine) of provenance of the sample. Overall, 14C calibration aims to express a raw 14C date plus/minus error (in years) in a probability density function (PDF) that quantifies the probability that the sample has a range of possible ages (in calendar years) given the amount of 14C present in layers of natural archives of known age (Keenan 2012).

\code{calendar} emulates the calibration protocol of the field-standard software \code{oxcal} (Bronk Ramsey 1995, 2007) using \code{\link[rcarbon]{calibrate}} (Crema & Bevan 2021) in F14C (radiocarbon concentration) space (Heaton et al. 2020a), and PDFs are normalized to a maximum probability of 1. The overarching goal is to provide a friendly tool for calibrating 14C dates, saving simple and tidy calibration outputs to the working directory, and computing statistical descriptors for examining PDF shape. The function employs one single computer core.

Calibration curves include \code{"intcal20"}, \code{"intcal13nhpine16"} or \code{"intcal13"} for fossils from the Northern Hemisphere (Reimer et al 2020), \code{"shcal20"}, \code{"shcal13shkauri16"} or \code{"shcal13"} for fossils from the Southern Hemisphere (Hogg et al 2020), and \code{"marine20"} or \code{"marine13"} for marine fossils (Heaton et al 2020b). \code{"intcal20"}, \code{"shcal20"} and \code{"marine20"} allow calibrations up to 55000 calendar years Before Present , and the other calibration curves allow calibrations up to 50000 calendar years Before Present.

Input data read from working directory using \code{\link[data.table]{fread}}; file extension and separators detected automatically. Data file must contain three or four columns. Columns 1 and 2 must always contain the raw radiocarbon dates and errors, respectively, and column 3 must always contain the dating-lab identifier (for instance, OxA-13579 or UCIAMS-123456). Use a 3-column data file in combination with \code{cal_curve} (for instance, \code{cal_curve = "intcal20"}), if all 14C dates should be calibrated with the same calibration curve. Use column four to include the name of calibration curves when different 14C dates should be calibrated with different calibration curves; argument \code{cal_curve} will be non-functional. Calibration results saved as a text file to working directory using \code{\link[data.table]{fwrite}}.

\code{calendar} computes three types of statistical descriptors, namely intercepts, moments and multi-modality:

INTERCEPTS: calibrated 14C dates expressed as the calibrated date at which a non-calibrated date intercepts a calibration curve (Telford, Heegaard and Birks 2003).
We have implemented the weighted average (mean of calibrated dates with non-zero probability weighted by their PDF probabilities), the median (calibrated date at which 50 percent of the PDF area is accounted for) and the mode (calibrated date at which the PDF probability is highest).
Confidence intervals of the weighted average (column \code{calW_sd} in output object \code{cal_intercepts}) can be plus/minus 1 standard deviation (\code{cal_ci = "sd1"} comprising 68.3 percent of the PDF, or 2 standard deviations (\code{cal_ci = "sd1"} comprising 95.4 percent of the pdf. For consistency with the weighted average, confidence intervals of the median (columns \code{calQ_lowerCI} and \code{calQ_upperCI} in output object \code{cal_intercepts}) can have quartile ranges from 0.1585 to 0.8415 (\code{cal_ci = "sd1"} comprising a cumulative 68.3 percent of the PDF area, or from 0.0230 to 0.9770 (\code{cal_ci = "sd2"} comprising a cumulative 95.4 percent of the PDF area.

MOMENTS: the shape of the PDF of a 14C date is summarized as the mean, variance, skewness (\code{\link[moments]{skewness}}) and kurtosis (\code{\link[moments]{kurtosis}}).

MULTI-MODALITY: the statistically supported modes (highest posterior density regions falling in the shortest intervals of the PDF) for 1 (\code{cal_ci = "sd1"} or 2 (\code{cal_ci = "sd2"} standard deviations using a (\code{summary} of a \code{\link[rcarbon]{calibrate}} object.
In output object \code{cal_modality}, column \code{modality_SD1} shows number of statistically supported modes at 1SD, and column \code{modality_SD2} shows number of statistically supported modes at 1SD and 2SD.
}
\examples{
#None

}
\references{
Bronk Ramsey, C. (1995) Radiocarbon calibration and analysis of stratigraphy: the OxCal program. Radiocarbon, 37, 425-430.\cr

Bronk Ramsey, C. (2017) Methods for summarizing radiocarbon datasets. Radiocarbon, 59, 1809-1833.\cr

Crema, E.R. & Bevan, A. (2021) Inference from large sets of radiocarbon dates: software and methods. Radiocarbon, 63, 23-39.\cr

Heaton, T.J. et al (2020a) The IntCal20 approach to radiocarbon calibration curve construction: a new methodology using Bayesian splines and errors-in-variables. Radiocarbon, 62, 821-863.

Heaton, T.J. et al (2020b) Marine20 - The marine radiocarbon age calibration curve (0-55000 cal BP). Radiocarbon, 62, 779-820.\cr

Hogg, A.G. et al (2020) SHCal20 Southern Hemisphere calibration, 0-55000 years cal BP. Radiocarbon, 62, 759-778.\cr

Keenan, D.J. (2012) Calibration of a radiocarbon age. Nonlinear Processes in Geophysics, 19, 345-350.\cr

Reimer, P.J. et al (2020) The IntCal20 Northern Hemisphere radiocarbon age calibration curve (0-55 cal kBP). Radiocarbon, 62, 725-757.\cr

Telford, R.J., Heegaard, E. & Birks, H.J.B. (2004) The intercept is a poor estimate of a calibrated radiocarbon age. Holocene, 14, 296-298.\cr
}
\seealso{
\code{\link[rcarbon]{calibrate}} (package \bold{rcarbon})
\code{\link[data.table]{data.table}} (package \bold{data.table})
\code{\link[data.table]{fread}} (package \bold{data.table})
\code{\link[data.table]{fwrite}} (package \bold{data.table})
\code{\link[moments]{skewness}} (package \bold{moments})
\code{\link[moments]{kurtosis}} (package \bold{moments})
}
