% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/griwm.R
\name{griwm}
\alias{griwm}
\title{Gaussian Re-sampled Inverse-Weighted McInerney (GRIWM)}
\arguments{
\item{chrono_data}{Text file read from working directory.
File must contain two columns without row names or column headings, column 1 includes dates, column 2 includes errors.
Unlimited number of dates (rows).
All dates must be Before Present, where Present = 1950.}

\item{signor_lipps}{Character. GRIWM accounts for the Signor-Lipps effect at the youngest (extinction) or oldest (arrival) end of a time series of dated observations.
Default \code{signor_lipps = "ext"} for extinction time, and \code{signor_lipps = "arr"} for arrival time.}

\item{biased}{Logical. Two estimators of extinction or arrival time.
Default is the original GRIWM estimator \code{biased = TRUE} dependent on parameter \code{alpha}.
Option \code{biased = FALSE} calculates an absolute time of extinction or arrival so is independent of parameter \code{alpha}. See details.}

\item{alpha}{Numeric. Sets the probability of finding a new observation younger than the youngest known observation (extinction) or older than the oldest known observation (arrival) of population or species.
Default \code{alpha = 0.05}. Lacks functionality for GRIWM unbiased estimator \code{biased = FALSE}. See details.}

\item{resample}{Numeric. Sets the number of samples taken from the normal distribution of each dated observation in the time series.
Default \code{resample = 10000}.}

\item{radiocarbon}{Character or numeric. Specifies if the time series contains radiocarbon dates.
Default \code{radiocarbon = "all"} assumes that all records in the time series are non-calibrated-radiocarbon dated observations.
\code{radiocarbon = 0} assumes that all records are in calendar years (whether dated by radiocarbon or not).
For data with mixture of non-calibrated radiocarbon dates and/or calibrated radiocarbon dates and/or non-radiocarbon dates, \code{radiocarbon} must be a vector of integers indicating the position of non-calibrated radiocarbon dates by row number. For instance \code{radiocarbon = c(2)} for one non-calibrated radiocarbon date in row 2, or \code{radiocarbon = c(2:4,7,11)} for five non-calibrated radiocarbon dates in rows 2 to 4, 7 and 11.}

\item{calibra}{Logical. Calibrates non-calibrated radiocarbon dates in calendar years.
Default \code{calibra = TRUE} assumes presence of non-calibrated radiocarbon dates if \code{radiocarbon = "all"} or \code{radiocarbon = c(row numbers)}.
See details.}

\item{cal_curve}{Character. Sets calibration curve for calibrating non-calibrated radiocarbon dates if present. See details.}

\item{upper14C}{Numeric. Sets the upper temporal boundary for radiocarbon calibrations.
Default \code{upper14C = 55000} for calibration curves \code{"intcal20"}, \code{"shcal20"} or \code{"marine20"}.
Use \code{upper14C = 50000} for calibration curves \code{"intcal13"}, \code{"shcal13"}, \code{"marine13"}, \code{"intcal13nhpine16"} or \code{"shcal13shkauri16"}}

\item{cal_save}{Logical. Determines if calibrated records are saved as text file to working directory.
Default \code{cal_save = "FALSE"}.}

\item{resample_save}{Logical. Determines if raw extinction (or arrival) times from each re-sampling iteration are saved as text file to working directory.
Default \code{resample_save = "FALSE"}.}

\item{griwm_save}{Logical. Determines if median extinction (or arrival) time and 95 percent CI are saved as text file to working directory.
Default \code{griwm_save = "FALSE"}.}
}
\value{
Returns a list containing two elements:
\itemize{
\item {\code{griwm_arguments}} {Data.frame with 2 columns.
 Column 1 = arguments in function \code{griwm}, Column 2 = values selected by user.}\cr
 \item {\code{calibration_BP}} {Data.table with 5 columns and as many rows as dates in the input data.
 Column 1 = non-calibrated dates, Column 2 = non-calibrated errors, Column 3 = calibrated dates, Column 4 = calibrated errors, Column 5 = type of date (radiocarbon, non-radiocarbon).
 Rows ordered from younger to older dates.}\cr
 \item {\code{griwm}} {Data.frame with 3 columns and 3 rows.
 Columns 1, 2 and 3 for 2.5 percent quartile CI, 50 percent (median) and upper 97.5 percent quartile CI of extinction (or arrival) time, respectively.
 Time in calendar years BCE (row 1: BCE = Before Current Era where CE starts at year 0), BP (row 2: BP = Before Present where Present = year 1950), BCY (row 3: BCY = Before Current Year).}
}
}
\description{
Estimate extinction time using a time series of dated observations with normal errors following Bradshaw et al (2012).
}
\details{
The fossil record is highly fragmented such that the youngest fossil of an extinct species is unlikely to belong to the last individual of that species (so-called Signor-Lipps effect by Raup 1986). Among the methods available for estimating extinction time (see Wang and Marshall 2016), GRIWM controls for the Signor-Lipps effect based on the frequency of dated fossils prior to the extinction event.
Analogously, GRIWM can also estimate arrival time by conceptualizing that the oldest fossil of a species recorded in a region is unlikely to belong to the first individual colonizing that region.
GRIWM accounts for dating uncertainty by means of a re-sampling procedure assuming a normal distribution of every fossil age in calendar years (Bradshaw et al 2012; Saltre et al 2015). Re-sampling implemented using using \code{\link[stats]{rnorm}}.

GRIWM defines extinction time as the calendar year at which the probability \code{alpha} of finding a fossil younger than the youngest known fossil is low (alpha = 0.05).
Analogously, GRIWM defines arrival time as the calendar year at which the probability \code{alpha} of finding a fossil older than the oldest known fossil is low (alpha = 0.05).
The method can be applied not only to fossil data but to any time series of historic observations each consisting of a date and an associated error.

\code{griwm} can compute biased and unbiased estimates of extinction (or arrival) time.
The biased estimator follows Bradshaw et al (2012) and generates extinction (or arrival) time given a pre-defined probability \code{alpha}.
The biased estimator generates an absolute extinction (or arrival) time independent of a pre-defined probability \code{alpha}.

\code{griwm} accepts any combination of chronological data (e.g., radiocarbon, ESR, OSL, TL, Uranium series).
Non-calibrated radiocarbon dates are calibrated in calendar years using \code{\link[rcarbon]{calibrate}} and expressed as a weighted average plus/minus a weighted standard deviation of the calibrated probability density function (Telford et al 2004).
Calibration curves include \code{"intcal20"}, \code{"intcal13nhpine16"} or \code{"intcal13"} for fossils from the Northern Hemisphere, \code{"shcal20"}, \code{"shcal13shkauri16"} or \code{"shcal13"} for fossils from the Southern Hemisphere, and \code{"marine20"} or \code{"marine13"} for marine fossils.
\code{"intcal20"}, \code{"shcal20"} and \code{"marine20"} allow calibrations up to 55000 calendar years Before Present, and the other calibration curves allow calibrations up to 50000 calendar years Before Present.
Dates from 1951 to present (post-bomb) not supported.
Default \code{cal_curve = "intcal20"}.

Input data read from working directory using \code{\link[data.table]{fread}}; file extension and separators detected automatically.
Output data saved to working directory using \code{\link[data.table]{fwrite}}.
\code{cal_save = TRUE} saves the original data with dates in years (non-calibrated) and calendar years (calibrated). Non-calibrated dates only included if \code{calibra = TRUE}. Name of output file = outputGRIWM_calibration.
\code{resample_save = TRUE} saves re-sampled extinction (or arrival) values used to estimate median extinction (or arrival) time with 95 percent quartile ranges. Number of values set by \code{resample}. Name of output file = outputGRIWM_resampled_Extinction (\code{signor_lipps = "ext"}) or outputGRIWM_resampled_Arrival (\code{signor_lipps = "arr"}).
\code{griwm_save = TRUE} saves median extinction (or arrival) time with 95 percent quartile ranges. Name of output file = outputGRIWM_medianCI_Extinction (\code{signor_lipps = "ext"}) or outputGRIWM_medianCI_Arrival (\code{signor_lipps = "arr"}).
}
\examples{
#None

}
\references{
Bradshaw, C.J.A., A. Cooper, C.S.M. Turney & B.W. Brook (2012) Robust estimates of extinction time in the geological record. Quaternary Science Reviews, 33, 14-19.\cr

Raup, D.M. (1986) Biological extinction in earth history. Science, 231, 1528-1533.

Saltre, F., B.W. Brook, M. Rodriguez-Rey, A. Cooper, C.N. Johnson, C.S.M. Turney & C.J.A Bradshaw (2015) Uncertainties in dating constrain model choice for inferring extinction time from fossil records. Quaternary Science Reviews, 112, 128-137.\cr

Telford, R.J., Heegaard, E. & Birks, H.J.B. (2004) The intercept is a poor estimate of a calibrated radiocarbon age. Holocene, 14, 296-298.

Wang, S.C. & C.R. Marshall (2016) Estimating times of extinction in the fossil record. Biology Letters, 12, 20150989.\cr
}
\seealso{
\code{\link[rcarbon]{calibrate}} (package \bold{rcarbon}),
\code{\link[data.table]{data.table}} (package \bold{data.table})
\code{\link[data.table]{fread}} (package \bold{data.table})
\code{\link[data.table]{fwrite}} (package \bold{data.table})
}
